+++ bin/actions/1010101-47.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-#: BL + V - Paste (Fire CTRL+V), when gnome screenshot active: export clipboard to image file and open it in XViewer
+#: BL + V - Paste (Fire CTRL+V), when gnome screenshot active: export copied image to file, than open in feh image viewer
 
 focusApplication=`cat /proc/$(xdotool getwindowpid $(xdotool getwindowfocus))/comm`
 
@@ -12,11 +12,8 @@ if [[ "$focusApplication" =~ "gnome-screensho" ]]; then  # gnome screenshot
   xdotool sleep 0.1
   xclip -selection clipboard -t image/png -o > /tmp/clipboard.png  # store image file
   xdotool sleep 0.2
-  xviewer /tmp/clipboard.png  # open image file in viewer
+  feh /tmp/clipboard.png  # open image file in viewer
 else  # fire regular paste
   xdotool sleep 0.1
   xdotool key ctrl+v
 fi
-
-
-